requirements.txt
asgiref==3.2.10
autopep8==1.5.3
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.0.8
django-ajax-selects==2.1.0
django-ckeditor==6.0.0
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-livereload-server==0.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.7.0
lockfile==0.12.2
Markdown==3.2.2
psycopg2==2.8.6
pycodestyle==2.6.0
python-slugify==4.0.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tornado==6.1
whitenoise==5.2.0
xlwt==1.3.0
zipp==3.1.0
